Algorithm Overhaul
Changes to main algorithm, changed coordinate system, bug fixes
when losing, typo fixes and additional comments for increased clarity.

-Algorithm change: Previous algorithm was prone to getting the AI into
unwinnable scenarios because it only focused on maximizing its own score
and not minimizing that of its opponent. This aggresive approach made it
easy to trap the AI because it barely made an attempt to thwart its
opponent.
-@Board.board now stored in a way that makes coordinates easier to
visualize (y,x). It also more accurately reflects locations of tokens
and nodes the board can now be represented as such:
[[(0,0),(0,1),(0,2)][(1,0),(1,1),(1,2)][(2,0),(2,1),(2,2)]]
-When losing, previous version would overwrite user values on the board
in order to win. This has since been fixed.
-Moved script to <head> for clarity. This was originally at the bottom
of the document to ensure that the script didn't run before the page
loaded. I've changed the JavaScript to use window.onload instead, and
moved the event listeners inside it. The rest of teh code should not
execute before it's called and should not cause problems in the same
way.
-Cursor changed to more clearly indicate a clickable object.
-Fixed minor typos
-Clarified comments