Added Buttons for turn decision
This adds some functionality to the game as well as appeases some of my
curiosity as to the different game possibilities for starting states. I
also removed some redundant code. It didn't seem important and didn't
break the code after testing, so I'm not sure why it was there in the
first place.

- Added buttons in the HTML to have the player decide whether they want
  to start the game or if they want the CPU to start
- Removed some redundant code
- Clarified and made minor modifications to old comments
- Updated README to reflect major changes.